make arguments as much readonly as possible
This promises users that we will leave their data structures alone, but
does not enforce users to make read only data structures before they pass
them to us.

We do return readonly types when we construct this ourselves, and type guards return assurances only for ReadOnly.

The point values can be objects, but we cannot make these readonly. ltCompare has to work for unknown, and unknown is not assignable to ReadOnly. This cascades up through the library.